+++ src/test/c_api_tests.cpp
@@ -386,7 +386,7 @@ TEST_F(CAPIInference, Basic) {
     uint32_t outputId = 0;
     OVMS_DataType datatype = (OVMS_DataType)199;
     const int64_t* shape{nullptr};
-    uint32_t dimCount = 42;
+    size_t dimCount = 42;
     OVMS_BufferType bufferType = (OVMS_BufferType)199;
     uint32_t deviceId = 42;
     const char* outputName{nullptr};
@@ -600,7 +600,7 @@ TEST_F(CAPIInference, ResponseRetrieval) {
     uint32_t outputId = 0;
     OVMS_DataType datatype = (OVMS_DataType)199;
     const int64_t* shape{nullptr};
-    uint32_t dimCount = 42;
+    size_t dimCount = 42;
     OVMS_BufferType bufferType = (OVMS_BufferType)199;
     uint32_t deviceId = 42;
     const char* outputName{nullptr};
@@ -668,7 +668,7 @@ protected:
     uint32_t outputId = 0;
     OVMS_DataType datatype = (OVMS_DataType)199;
     const int64_t* shape{nullptr};
-    uint32_t dimCount = 42;
+    size_t dimCount = 42;
     OVMS_BufferType bufferType = (OVMS_BufferType)199;
     uint32_t deviceId = 42;
     const char* outputName{nullptr};
